Skips removed subjects

+++ code/python/organise_beh.py
@@ -18,8 +18,13 @@ if __name__ == '__main__':
             if beh_dir.exists() and beh_dir.is_dir():
                 for beh_file in beh_dir.iterdir():
                     if beh_file.is_file():
-                        target_dir = Path(raw_path) / subject_dir.name / Path('beh')
-                        target_file = target_dir / beh_file.name
+                        target_subject_dir = Path(raw_path) / subject_dir.name
+                        if target_subject_dir.exists() and target_subject_dir.is_dir():
+                            target_dir = target_subject_dir / Path('beh')
+                            target_file = target_dir / beh_file.name
 
-                        target_dir.mkdir(parents=False, exist_ok=True)
-                        target_file.write_bytes(beh_file.read_bytes())
+                            target_dir.mkdir(parents=False, exist_ok=True)
+                            target_file.write_bytes(beh_file.read_bytes())
+                        else:
+                            print(f"Directory {target_subject_dir} does not exist")
+                            break
